refactor: move dependecy check out of `run_app`
The first step in `run_app` is to check if necessary external
dependencies are installed. This does not need to be a method of the
`App` struct nor does it need to part of `run_app`. Move dependecy check
out of `run_app`.

In the future, this check can be made by injecting any necessary
configuration information instead of it being a method on the `App` struct.

Signed-off-by: Ivin Joel Abraham <ivinjabraham@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>