style(patch-hub-proc-macros): rename crate `proc-macros` for `patch-hub-proc-macros`
Following the effort to publish `patch-hub` into crates.io, we need to
also publish the crate called `proc-macros`, which, at moment, only
contains a procedural macro to cleanly generate the default
deserialization values of XMLs that have missing fields.

As such, having a crate named `proc-macros` doesn't seem fit, so rename
it `patch-hub-proc-macros` to tie it to its "parent" project.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>