test(patch_renderer): remove flaky test suite
The `app::patch_renderer` module is really important to bring life for
the patch previews in `patch-hub`. Initially, I added a test suite for
it, which is always a good idea, and it seemed reasonable, as we could
have a `String` type with the expected and check if the renderer
produced the same. However, this test implementation was really coupled
with the setup where the test being run, which made it flaky.

Flaky tests don't help in nothing, and just occupy space, so remove this
test suite completely.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>